requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
tensorflow==2.15.0
